requirements.txt
Flask==3.1.0
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
qrcode==8.0
Pillow==11.0.0
openpyxl==3.1.5
